Update Mode For ImageHost

+++ src/main/java/cn/mintimate/filecloudplus/entity/UserIp.java
@@ -36,7 +36,7 @@ public class UserIp implements Serializable {
 
     private String targetClassify;
 
-    private Integer targetId;
+    private Long targetId;
 
 
 }
